functions/requirements.txt
# requirements.txt
firebase-functions>=0.2.0
firebase-admin>=6.3.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth>=2.23.3